feat: support wasm build (#673)
